refactor: move useTaxonomyDetail hooks to avoid circular dependency
manage-orgs -> taxonomy-detail -> taxonomy-menu -> manage-orgs
