+++ src/taxonomy/data/api.test.js
@@ -10,6 +10,7 @@ import {
   getTaxonomyListData,
   getTaxonomyExportFile,
   getTaxonomyApiUrl,
+  getTaxonomy,
   deleteTaxonomy,
 } from './api';
 
@@ -66,6 +67,13 @@ describe('taxonomy api calls', () => {
     await deleteTaxonomy();
 
     expect(axiosMock.history.delete[0].url).toEqual(getTaxonomyApiUrl());
+  });
+
+  it('should call get taxonomy', async () => {
+    axiosMock.onGet(getTaxonomyApiUrl(1)).reply(200);
+    await getTaxonomy(1);
+
+    expect(axiosMock.history.get[0].url).toEqual(getTaxonomyApiUrl(1));
   });
 
   it('should set window.location.href correctly', () => {
