+++ src/taxonomy/data/apiHooks.test.jsx
@@ -1,9 +1,12 @@
 import { useQuery, useMutation } from '@tanstack/react-query';
 import { act } from '@testing-library/react';
+
 import {
   useTaxonomyListDataResponse,
   useIsTaxonomyListDataLoaded,
   useDeleteTaxonomy,
+  useTaxonomyDetailDataStatus,
+  useTaxonomyDetailDataResponse,
 } from './apiHooks';
 import { deleteTaxonomy } from './api';
 
@@ -73,5 +76,40 @@ describe('useDeleteTaxonomy', () => {
       await mutationFn({ pk: 1 });
       expect(deleteTaxonomy).toBeCalledWith(1);
     });
+  });
+});
+
+describe('useTaxonomyDetailDataStatus', () => {
+  it('should return status values', () => {
+    const status = {
+      isError: false,
+      error: undefined,
+      isFetched: true,
+      isSuccess: true,
+    };
+
+    useQuery.mockReturnValueOnce(status);
+
+    const result = useTaxonomyDetailDataStatus(0);
+
+    expect(result).toEqual(status);
+  });
+});
+
+describe('useTaxonomyDetailDataResponse', () => {
+  it('should return data when status is success', () => {
+    useQuery.mockReturnValueOnce({ isSuccess: true, data: 'data' });
+
+    const result = useTaxonomyDetailDataResponse();
+
+    expect(result).toEqual('data');
+  });
+
+  it('should return undefined when status is not success', () => {
+    useQuery.mockReturnValueOnce({ isSuccess: false });
+
+    const result = useTaxonomyDetailDataResponse();
+
+    expect(result).toBeUndefined();
   });
 });
