--- src/taxonomy/taxonomy-detail/data/types.mjs
@@ -1,22 +1,0 @@
-// @ts-check
-
-/**
- * @typedef {Object} TaxonomyData
- * @property {number} id
- * @property {string} name
- * @property {string} description
- * @property {boolean} enabled
- * @property {boolean} allowMultiple
- * @property {boolean} allowFreeText
- * @property {boolean} systemDefined
- * @property {boolean} visibleToAuthors
- * @property {string[]} orgs
- * @property {boolean} allOrgs
- * @property {number} tagsCount
- */
-
-/**
- * @typedef {Object} UseQueryResult
- * @property {Object} data
- * @property {string} status
- */
