+++ src/taxonomy/taxonomy-detail/index.js
@@ -1,3 +1,2 @@
 // @ts-check
-export { default as TaxonomyDetailPage } from './TaxonomyDetailPage';
-export { useTaxonomyDetailDataResponse, useTaxonomyDetailDataStatus } from './data/apiHooks';
+export { default as TaxonomyDetailPage } from './TaxonomyDetailPage'; // eslint-disable-line import/prefer-default-export
